@@ -9,7 +9,7 @@
     "description": "Project contains boilerplate that allows you to start writing web apps with using Typescript, React, Redux, Rect Router v5, webpack(HMR), express and gulp",
     "main": "index.js",
     "engines": {
-        "node": "^16.15.0"
+        "node": "^18.16.0"
     },
     "scripts": {
         "eslint": "cross-env NODE_ENV=production gulp eslint",
@@ -79,45 +78,47 @@
         "core-js": "3.23.4",
         "cross-env": "7.0.3",
         "css-loader": "6.7.1",
-        "eslint": "8.20.0",
+        "eslint": "^8.52.0",
         "express": "4.18.1",
         "file-loader": "6.2.0",
         "fs-extra": "10.1.0",
-        "graceful-fs": "4.2.10",
-        "gulp": "4.0.2",
+        "graceful-fs": "^4.2.11",
+        "gulp": "^4.0.2",
         "gulp-clean": "0.4.0",
-        "gulp-debug": "4.0.0",
+        "gulp-debug": "^4.0.0",
         "html-webpack-plugin": "5.5.0",
         "jest": "28.1.3",
         "jest-cli": "28.1.3",
-        "jest-runner-eslint": "1.0.1",
-        "jest-runner-stylelint": "2.3.7",
+        "jest-runner-eslint": "^2.1.2",
+        "jest-runner-stylelint": "^2.3.7",
         "less": "4.1.3",
         "less-loader": "11.0.0",
         "lodash": "4.17.21",
         "mini-css-extract-plugin": "2.6.1",
         "open": "8.4.0",
-        "prettier": "2.7.1",
-        "react-hot-loader": "^4.13.0",
+        "prettier": "^3.0.3",
         "resolve-url-loader": "5.0.0",
         "run-sequence": "2.2.1",
         "source-map-loader": "4.0.0",
         "style-loader": "3.3.1",
-        "stylelint": "14.9.1",
+        "stylelint": "^15.11.0",
         "svg-url-loader": "7.1.1",
         "ts-jest": "28.0.7",
         "ts-loader": "9.3.1",
         "ts-node": "10.9.1",
         "tsconfig-paths-webpack-plugin": "3.5.2",
-        "typescript": "4.7.4",
-        "webpack": "5.73.0",
-        "webpack-dev-middleware": "5.3.3",
-        "webpack-dev-server": "4.9.3",
-        "webpack-env": "0.8.0",
-        "webpack-hot-middleware": "2.25.1",
-        "webpack-merge": "5.8.0",
-        "webpack-sources": "3.2.3",
-        "webpack-typings-for-css": "0.9.0",
-        "yargs": "17.5.1"
-    }
+        "type-fest": "^4.5.0",
+        "typescript": "5.2.2",
+        "typescript-plugin-styled-components": "^3.0.0",
+        "webpack": "^5.89.0",
+        "webpack-dev-middleware": "^6.1.1",
+        "webpack-dev-server": "^4.15.1",
+        "webpack-env": "^0.8.0",
+        "webpack-hot-middleware": "^2.25.4",
+        "webpack-merge": "^5.10.0",
+        "webpack-sources": "^3.2.3",
+        "webpack-typings-for-css": "^0.12.0",
+        "yargs": "17.7.2"
+    },
+    "packageManager": "yarn@1.22.19"
 }
