@@ -1,8 +1,8 @@
-import * as ReactDOM from "react-dom";
-import { AppContainer } from "react-hot-loader";
-import { App } from "./App";
+import { createRoot } from "react-dom/client";
+import App from "./App";
 import "core-js/stable";
 import "regenerator-runtime/runtime"
+import { Greeter } from 'my-awesome-greeter';
 
 import { configuration } from "../build/configuration";
 
@@ -16,24 +16,28 @@ declare const require: {
 
 declare const module: { hot: any };
 
-const render = (Component) => {
-    ReactDOM.render(
-        <AppContainer>
-            <Component/>
-        </AppContainer>,
-        document.getElementById(configuration.mountPoint)
-    );
+const container = document.getElementById(configuration.mountPoint);
+if(!container) {
+    throw new Error(`No container with id ${configuration.mountPoint}`);
+}
+
+const root = createRoot(container)
+
+
+const render = (rootElement, Component) => {
+    Greeter("Maksym");
+    rootElement.render(<Component />);
 };
 
 window.onload = () => {
-    render(App);
+    render(root, App);
 };
 
 // Hot Module Replacement API
 if (module.hot) {
     module.hot.accept('./App', () => {
         /* eslint-disable @typescript-eslint/no-var-requires */
-        const NextApp = require('./App');
-        render(NextApp);
+        const NextApp = require('./App').default;
+        render(root, NextApp);
     });
 }
