@@ -1,22 +1,15 @@
 /* Styles */
 import './styles.less';
+import { useRootApplication } from "@dhampir/core";
 
-/* Routing */
 import './extensions';
 
-// import { useRootApplication, StorageType } from '@dhampir/core';
-import { FunctionComponent } from 'react';
-import {BrowserRouter} from "react-router-dom";
-import {Layout} from "@components/layout";
-import {Route, Routes} from "react-router";
+const App = () => {
+    const Application = useRootApplication({
+        storageType: "redux"
+    });
 
-export const App: FunctionComponent = () => {
-    // const Application = useRootApplication({
-    //     storageType: StorageType.REDUX
-    // });
-
-    //return Application !== undefined ? <Application.Component {...Application.props} /> : null;
-    return <BrowserRouter>
-        <Routes><Route path={'/'} element={<Layout />} /></Routes>
-    </BrowserRouter>
+    return Application !== undefined ? <Application.Component {...Application.props} /> : null;
 };
+
+export { App as default }
