@@ -8,7 +8,6 @@ import {
     RoutingArea,
     Row, Units,
     Spacer,
-    areAreasVisible,
     BorderSide
 } from '@dhampir/core';
 
@@ -22,12 +21,15 @@ export const Footer: FunctionComponent = () => {
     return <Row>
             <Decorator fill={true} borderPosition={[BorderSide.TOP]} borderWidth={1}>
                 <Spacer space={0.5} units={Units.EM}>
-                    <Routes>
-                        <Route path={'/one'} element={<div>One</div>} />
-                        <Route path={'/two'} element={<div>Two</div>} />
-                        <Route path={'/three'} element={<div>Three</div>} />
-                        <Route path={'/four'} element={<div>Four</div>} />
-                    </Routes>
+                    {isAreaVisible(BOTTOM_LEFT, location.pathname) && <Column>
+                        <Area area={BOTTOM_LEFT} />
+                    </Column>}
+                    {isAreaVisible(BOTTOM_CENTER, location.pathname) && <Column greedy={true}>
+                        <Area area={BOTTOM_CENTER} />
+                    </Column>}
+                    {isAreaVisible(BOTTOM_RIGHT, location.pathname) && <Column>
+                        <Area area={BOTTOM_RIGHT} />
+                    </Column>}
                 </Spacer>
             </Decorator>
         </Row>;
